Added modal component Confirm, added editUser page

+++ src/components/usersTable/UsersTable.tsx
@@ -1,15 +1,20 @@
 'use client'
 
+import axios from 'axios'
+import { CiSearch } from 'react-icons/ci'
+import 'react-toastify/ReactToastify.min.css'
 import { MdDeleteForever } from 'react-icons/md'
+import { IoCloseOutline } from 'react-icons/io5'
+import { HiArrowsUpDown } from 'react-icons/hi2'
+import { FaUserEdit } from 'react-icons/fa'
+import { useEffect, useRef, useState } from 'react'
+import { ToastContainer, toast } from 'react-toastify'
+
 import { User } from '@/types'
 import styles from './usersTable.module.css'
-import axios from 'axios'
-import { ToastContainer, toast } from 'react-toastify'
-import 'react-toastify/ReactToastify.min.css'
 import { revalidateByPath } from '@/actions/actions'
-import { useEffect, useRef, useState } from 'react'
-import { CiSearch } from 'react-icons/ci'
-import { IoCloseOutline } from 'react-icons/io5'
+import ConfirmModal from '../UI/confirmModal/ConfirmModal'
+import { usePathname, useRouter } from 'next/navigation'
 
 type Props = {
 	users: User[]
@@ -29,36 +34,56 @@ export default function UsersTable({ users }: Props) {
 	const [isReverseSorting, setIsReverseSorting] = useState<boolean | undefined>(
 		false
 	)
+	const [soringKey, setSortingKey] = useState<string | undefined>('')
 	const [filteredUsers, setFilteredUsers] = useState<User[] | undefined>([])
 	const [isInputValueLength, setIsInputValueLength] = useState<
 		boolean | undefined
 	>(false)
 	const [searchValue, setSearchValue] = useState<string | undefined>('')
-	const searchRef = useRef<HTMLInputElement>(null)
+	const [confirmDelete, setConfirmDElete] = useState<boolean | undefined>(false)
+	const [deleteUserId, setDeleteUserId] = useState<string | undefined>('')
+	const dialogRef = useRef<HTMLDialogElement>(null)
+	const router = useRouter()
+	const pathName = usePathname()
 
 	useEffect(() => {
 		if (sortedUsers!.length < 1) setSortedUsers([...users])
 	}, [searchValue])
+
 	useEffect(() => {
 		setFilteredUsers([...users!])
 	}, [])
+
+	useEffect(() => {
+		if (confirmDelete) {
+			dialogRef.current?.showModal()
+		} else {
+			dialogRef.current?.close()
+		}
+	}, [confirmDelete])
+
 	const onDeleteUserHandler = async (userId: string) => {
-		try {
-			const res = await axios.delete('http://localhost:3000/api/users', {
-				data: {
-					id: userId
+		console.log('delete')
+
+		if (deleteUserId) {
+			try {
+				const res = await axios.delete('http://localhost:3000/api/users', {
+					data: {
+						id: userId
+					}
+				})
+				if (res.status) {
+					toast.success(res.data.message, { position: 'top-right' })
+					revalidateByPath('/profile/users')
 				}
-			})
-			if (res.status) {
-				toast.success(res.data.message, { position: 'top-right' })
-				revalidateByPath('/profile/users')
+			} catch (error: any) {
+				console.log(error)
 			}
-		} catch (error: any) {
-			console.log(error)
 		}
 	}
 
 	const onSortingHandler = (key: string) => {
+		setSortingKey(key)
 		if (isReverseSorting) {
 			const sorted = sortedUsers?.sort((a: any, b: any) => {
 				return a[key] < b[key] ? 1 : a[key] > b[key] ? -1 : 0
@@ -73,6 +98,7 @@ export default function UsersTable({ users }: Props) {
 			setIsReverseSorting(!isReverseSorting)
 		}
 	}
+
 	const onChangeHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
 		setSearchValue(e.target.value)
 		e.target.value.length > 0
@@ -85,10 +111,17 @@ export default function UsersTable({ users }: Props) {
 		setFilteredUsers([...filtered])
 	}
 	const onResetSearchValueHandler = () => {
+		setFilteredUsers([...users])
 		setIsInputValueLength(false)
 		setSearchValue('')
 	}
-
+	const onResetHandler = () => {
+		setConfirmDElete(false)
+	}
+	const onConfirmDeleteHandler = (userId: string) => {
+		setDeleteUserId(userId)
+		setConfirmDElete(true)
+	}
 	return (
 		<div className={styles.usersTableContainer}>
 			<ToastContainer />
@@ -99,7 +132,6 @@ export default function UsersTable({ users }: Props) {
 					placeholder="Пошук..."
 					onChange={onChangeHandler}
 					value={searchValue}
-					ref={searchRef}
 				/>
 				{isInputValueLength ? (
 					<IoCloseOutline
@@ -121,7 +153,14 @@ export default function UsersTable({ users }: Props) {
 								key={item.name}
 								onClick={() => onSortingHandler(item.name)}
 							>
-								{item.title}
+								<div className={styles.usersTableThHeaedContainer}>
+									{item.title}
+									{soringKey === item.name ? (
+										<HiArrowsUpDown size={14} className={styles.sortingIcon} />
+									) : (
+										<></>
+									)}
+								</div>
 							</th>
 						))}
 					</tr>
@@ -143,19 +182,40 @@ export default function UsersTable({ users }: Props) {
 								<td
 									className={`${styles.usersTableTd} ${styles.usersTableTdBtn}`}
 								>
-									<div className={styles.deleteUserContainer}>
-										<MdDeleteForever
-											size={24}
-											className={styles.deleteIcon}
-											onClick={() => onDeleteUserHandler(user?.id!)}
-										/>
-										<span className={styles.toolTip}>Видалити</span>
+									<div className={styles.userTableActionsContainer}>
+										<div className={styles.deleteUserContainer}>
+											<MdDeleteForever
+												size={24}
+												className={styles.deleteIcon}
+												onClick={() => onConfirmDeleteHandler(user.id!)}
+											/>
+											<p className={styles.deleteUserToolTip}>Видалити</p>
+										</div>
+										<div className={styles.editUserContainer}>
+											<FaUserEdit
+												size={24}
+												className={styles.editUserIcon}
+												onClick={() => router.push(`${pathName}/${user.id}`)}
+											/>
+											<p className={styles.editUserToolTip}>Редагувати</p>
+										</div>
 									</div>
 								</td>
 							</tr>
 						))}
 				</tbody>
 			</table>
+			<dialog ref={dialogRef} className={styles.userDeleteDialog}>
+				<form method="dialog">
+					{confirmDelete && (
+						<ConfirmModal
+							userId={deleteUserId!}
+							onConfirmHandler={() => onDeleteUserHandler(deleteUserId!)}
+							onResetHandler={onResetHandler}
+						/>
+					)}
+				</form>
+			</dialog>
 		</div>
 	)
 }
